Add explicit casts for 64-bit compatibility
Add explicit type casts throughout container and utility classes to resolve
compiler warnings about implicit conversions between size_t, pointer
differences, and int types. Changes include:

- Pointer arithmetic to int casts in XArray, XClassArray, XSArray
- size_t to int casts in FixedSizeAllocator and VxConfiguration
- strlen result casts in VxConfiguration and XString
- Improved boolean return patterns replacing C-style casts to XBOOL

These changes ensure correct behavior on 64-bit platforms where size_t
differs in size from int.

+++ include/FixedSizeAllocator.h
@@ -188,7 +188,7 @@ private:
                     size_t marked = 0;
                     while (marked < m_BlockAvailables && freeb != INVALID_BLOCK_INDEX)
                     {
-                        freeBlocks.Set(freeb);
+                        freeBlocks.Set((int)freeb);
                         unsigned char *p = m_Data + freeb * iBlockSize;
                         freeb = ReadFreeListIndex(p);
                         ++marked;
@@ -200,7 +200,7 @@ private:
                     for (size_t i = 0; i < iBlockCount; ++i)
                     {
 
-                        if (freeBlocks.IsSet(i))
+                        if (freeBlocks.IsSet((int)i))
                             continue;
 
                         unsigned char *p = m_Data + i * iBlockSize;
